retrySaveMemberId() 위치조정
The problem is that you are trying to save the member ID to local storage before you have received it from the server. The retrySaveMemberId() function is called when the request to the server is successful, but the locationHeader parameter is not set until the response body has been received. You can fix this by moving the call to retrySaveMemberId() inside the .then() callback for the fetch() request.